requirements.txt
python-telegram-bot>=21.0
httpx>=0.27.0
pyyaml>=6.0
